exercise 4.21

+++ part-04/bloglist/controllers/blogs.js
@@ -32,6 +32,30 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  if (!user) {
+    return response
+      .status(400)
+      .json({ error: `user ${decodedToken.username} does not exist anymore` });
+  }
+
+  if (!blog) {
+    return response.status(404).end();
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response
+      .status(401)
+      .json({ error: `user ${user.username} is not the creator of this blog` });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
